eslint-config,eslint-plugin-custom: Move local rules to custom plugin
We're doing this because it fixes ESLint in VS Code: the VS Code extension is incompatible with the eslint local rules plugin. This makes the dev experience much better as devs then can see the ESLint issues as they go.

The behaviour before and after is identical in terms of what lint errors are thrown.
