+++ libraries/eslint-plugin-custom/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@bluedot/eslint-plugin-custom",
+  "version": "0.0.0",
+  "private": true,
+  "main": "src/index.js"
+}
