+++ libraries/eslint-plugin-custom/src/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  rules: {
+    'no-default-tailwind-tokens': require('./rules/no-default-tailwind-tokens'),
+  },
+};
